refactor: 스레드 태그의 이름명 변경
